Fix usage of RMW-intrinsics in core.atomic.atomicOp()
The operand needs to be cast to the type T in memory.

When repeating the operation for the atomicOp() return value, which is
inconveniently defined as the final result, I'm not sure whether we really get
the same result.
The LLVM intrinsic will compute the operation using type T, while D e.g. refuses
to add 2 ubytes (it casts both operands to int before adding).
This also means that we need to cast the 2nd operation's result to T, although
both operands are of type T.