Base implicit executable filename on first object file name
And only then on an optional root module name.
This should mimic DMD behavior (I haven't checked its source though)
and so makes sure LDMD is compatible in this regard.

E.g., when compiling a module `a.d` stating `module tst;` to an executable
without providing an explicit output file name, the executable will now be
called `a[.exe]` instead of `tst[.exe]`.