Properly fix Issue 15296 - expand CallExp in ExpStatement as statements
If a `CallExp`, `CondExp`, or `CommaExp` appears in `ExpStatement`, it can be inlined as statements.
It provides more inlining opportunity. Even if the called function contains some statements which cannot be expanded as expressions (e.g. `ForStatement`), the function call can be inlined.

In `expandInline`, if `vthis` is a temporary variable, its dtor call should be deferred till the end of expanded function body statements.