Extend DtoBitCast() to support aggregate values.
The LLVM instruction bitcast works only on non-aggregate values and types.
As long as struct are stored in memory there is no problem. But if the
struct has integral size (e.g. only an int member) then the value is not
necessarily stored in memory. Nevertheless there may be the need to cast
this value to a different type.
DtoBitCast() is extended to store and reload the value with differnt types
in this case.

This functionality is used to fix issue #1211.