Replace UTF-8–invalid embedded characters with valid UTF-8
The copyright symbol, when encoded in Win-1252, is not valid UTF-8. D assumes UTF-8 and chokes on it, so this commit replaces it with the valid, multibyte UTF-8 encoding of the symbol.

Note: I feel it would be safer to simply use "(C)", as is used everywhere else in the code, but that would seem to ignore the intentions of the original author.