Emit symbols by default as COMDATs
That's apparently what DMD does and what currently happens for all
non-MSVC targets anyway.

For MSVC, we need this to enforce the `any` selection instead of
`no duplicates`, otherwise the linker complains about multiply defined
symbols and refuses to link without an extra command line option
(which in turn outputs a non-suppressable warning).