Use special ref var machinery for CondExp temporary
The semantics (hidden pointer to values) are the same, so it makes
sense to reuse the associated code. Delaying the load to wherever
the variable is actually used might make for somewhat more efficient
codegen.

Incidentally, this fixes GitHub #1294, but that is more a case
of the underlying bug not being triggered anymore by that
particular test case.