port: Use std::* versions cmath/ctype functions
The global namespace versions are not guaranteed to be available in
C++ mode, and GCC 5.3 on Arch Linux indeed does not make them available.

This is just long-term maintenance (the implementation has been
moved to D in DDMD/LDC master), so no IN_LLVM business anymore.

For now, this just changes the "generic" Posix branch, but eventually
we might want to move the other OSes to using std too.

GitHub: Fixes #1395.