Use build host D compiler real type
D needs to compile itself on non-x86 hosts where real type might be
something other than 80-bit extended precision.  The simplest approach
is to use the existing D compiler real type as it must already have
correct real.nan and snan intializer.

Note this does not work for cross-compiling to alternate targets and is
only intended as transitional until that capability is added.