Check existing runtime function declarations for matching type
This fixes issue #958 by checking the function type of an existing
function declaration with the same runtime function name inside the
codegen'd module.
This should normally be an existing forward declaration from a previous
runtime call, but the user is free to declare (and even implement) it
too.

The LOC for the error is empty as the provided `loc` may likely be from
a template instantiation from another module.