DebugCondition: Introduce D overload and document
This patch introduces overload that accept D strings,
and make the C++ version forward to them.

The C++ overload used outside DMD are marked as deprecated
in order to avoid new usage of them within DMD, but should be
kept for C++ drivers (LDC, GDC).

Documentation was also added to the static methods and the ctor.