Correctly handle arrdspaces (AS) other than zero
LDC currently assumes that there is only one addrspace, zero.

dcompute requires the use of multiple AS to deal with global/shared
generic pointer.

There is only one use (bar asserts) that I have identified, (there may
be more, but libmir/dcompute build fine with #1786 HEAD~3) which is
generating an invalid  bitcast, because the types are not equal, but
for the purposes of copying do not require, and cannot be done with, a
bitcast.