[NFC] Refactor emission of debug info.
Better annotate reasons for not emitting debug info. Also make it
easier to change conditions under which DI should be emitted.

This will make it easier when more dcompute related code is added, as
there cannot be any DI emission for dcompute (SPIRV does allow some
extra debug info but that is not a target of dcompute yet.