Ddoc: add new macros: DDOC_MEMBER_HEADER and DDOC_HEADER_ANCHOR
The second parameter ($1) of the DDOC_HEADER_ANCHOR macro expands to
the fully qualified name of the current symbol without the package and
module prefix. That is, for a module level symbol it's the name of a
symbol. For a nested symbol, like a method in a class, it's the symbol
name prefixed with the class name. The third parameter ($2) expands to
just the name of the current symbol.

The DDOC_MEMBER_HEADER wraps DDOC_HEADER_ANCHOR to allow a sensible
output format. The DDOC_MEMBER_HEADER is wrapped in the DDOC_MEMBER
macro.

By default these two macros expand to nothing, to keep backwards
compatibility.