Print exception and abort if no handler is found
Previously, dwarf exception handler was relying on a fact that running
user code is always enclosed in the catch-all exception handler.  That
might not be the case if the user (with debugger, for example) change
value of `rt_trapExceptions` to false. To cover this scenario, instead
of issuing fatal error, we wil just print exception info and abort the
program.