CMake: Tweak quotes around cmdline options
E.g., LDC options such as `-L"-Wl,..."` don't work for Travis OSX, so use
`"-L-Wl,..."` instead. This may also help with expanding the flags, as
they aren't a CMake list (a string with semicolons separating the items),
but a string with the arguments separated by spaces (we do use
`separate_arguments()` in some places later on to convert it to a list).