[dcompute] target and codegen part 1
Adds the command line option `-mdcompute-targets=` a comma separated
list of targets to generate for dcompute (cl_options.\*), and hides
options added by the spirv backend.

Adds an analogue to `CodeGenerator` (driver/dcomputecodegenerator.\*)
that deputises to the targets (gen/dcomputetarget\*) to do the code
generation and add required metadata.

Adds code to detect types from `ldc.dcomputetypes`, a module for magic
types required for targeting SPIRV and NVPTX. Currently only `Pointer`
a struct containing a pointer to a specified address space mapped to
the correct addrspace for the given backend (gen/dcomputetypes.*\).