Add option to avoid recursing in get_git_head_revision cmake function
Since this function is only ever invoked from the base directory of the
LDC source tree, it does not need to recurse in order to get the correct
git revision for LDC.  This patch therefore adds a parameter that can be
set to `TRUE` or `FALSE` depending on whether one wants to use recursion
or not, and sets it to false in when calling `get_git_head_revision` in
LDC's own `CMakeLists.txt`.

This should prevent issues that arise when building from source tarballs
that have been extracted into a subdirectory of another git tree (!).
This is unfortunately a real possibility: apart from folks using git to
version their home directory [1], it has also proven an issue when using
source tarballs to build the LDC snap package [2].

[1]https://stackoverflow.com/questions/21266948/use-git-to-manage-home-directory/42946460
[2]https://github.com/ldc-developers/ldc2.snap/issues/21#issuecomment-287575769

Fixes https://github.com/ldc-developers/ldc/issues/2042.