Remove ASTNull
After reviewing this in more detail the conclusion is that there is no use case for the ASTNull AST family that is not better covered by ASTBase and therefore should be removed.

While this does cover the API used by the parser to construct an AST it does not actually result in building an AST, and at the same time is ineffective for use cases that don't require an AST as it will allocate huge amounts of memory and does not provide a hookable API.  The testing of the parser API will be adequately covered by ASTBase.