Add bin/build-ldc-runtime.sh helper script
This new script will automatically download and extract the full LDC
source archive of the release it was packaged with and then build
druntime/Phobos stand-alone (taking a few minutes) using

* user-supplied CMake command-line args,
* the LDC compiler in the same directory as the script for the D parts,
* and the `CC` C compiler for the C/ASM parts and linking the shared libs
  (on POSIX at least).

So we can build on top of that to let the user build the runtime libs with
LTO support or for cross-compilation targets any time he/she wants to.

Requirements are CMake >= 2.8.9 and make [as well as bash, wget and tar]
in addition to an anyway-required C toolchain.