Travis Linux: Try to work around memory limitations
The workers may have less than 4 GB of memory and some jobs seem to get
killed due to this, particularly when building the stdlib test runners.

So try precompiling the shared unittest objects manually with `-j3` too.
If compiled as part of the testrunner tests, ninja apparently defaults to
a high number of parallel jobs based on the CPU and not on the available
CPU cores for the Docker container.