Runtime CMake: Make sure test runners are re-linked
add_dependencies() only makes sure the dependencies are up-to-date before
running the testrunner's target. So after hacking a source file, the
unittest libraries were correctly updated, but the test runners themselves
weren't relinked afterwards.
target_link_libraries() isn't an option when linking in the library
manually via custom linker flags to make sure all objects are pulled in.

Fortunately, the LINK_DEPENDS target property allows specifying a file
dependency, so that a testrunner is relinked when its library is newer.