Fix DMD issue 17502
Apparently the front-end ends up with nestedrefs for a function's
special __result variable if the function's return type needs to be
inferred (auto) and the variable is referenced in a parameter-less out
contract.
The variable is not really nested though; it's passed by reference to
the out contract function, as if it was declared with a parameter:
`out (__result) { ... }`.

Fixes compilable/test17502.