Fix bug 17502 (again): Generate contracts after inferring return type.
Wait until after the return type has been inferred before generating the
contracts for this function, and merging contracts from overrides.

This was originally at the end of the first semantic pass, but required
a fix-up to be done here for the '__result' variable type of __ensure()
inside auto functions, but this didn't work if the out parameter was
implicit.