[Cherry-pick] rt.config: Make custom mangling more robust
dlang/druntime@49f4a91952141ddf55c054829026dd6c048609cc

`pragma(mangle, ...)` allows specifying a 'base' mangle for the symbol's
linkage. So by using `extern(C) pragma(mangle, ...)`, a base C mangle can
be specified, without having to take care of a platform-specific prefix;
that is handled by the compiler.

The previous code relied on D symbols not getting prefixed, which isn't
the case for LDC on OSX [and Win32, but that's subject to change].