sections_elf_shared: Actually print error message when trying to use shared libraries with static runtime (#2454)
Using assert() in this module is problematic in more than one way:
In debug mode, most of the assertion failures will come at a point
where the runtime is not sufficiently initialised to properly throw
the AssertError yet. On the other hand, in release mode, the assert(0)s
will generate traps that are unhelpful for users.

A patch converting the other assertions will be proposed upstream.