Improve test suite interface.
Makes the test suite a little more accomodating to developers who don't know how to use it yet. It now indicatest that the suite is meant to be run through the makefile (not by running d_do_test.d directory) and also prints a nice error message when environment variables are missing that would have been set by the Makefile.  This is a nice contrast to what was done before which would have been to fail at some later point in the test with a cryptic error message as a result of some environment variable being empty.