Refactor Objective-C glue layer
This is a followup on the refactoring of the frontend part. This moves
the interface of the Objective-C glue layer to a class to make it
possible to determine at runtime (of the compiler) if Objective-C is
supported or not. This allows cross-compiling and the
`objc_glue_stubs.d` file can be removed.