Upstream GDC patch to mangle va_list as a pointer.
The default behaviour is to raise an internal compiler error and exit if
cppmangle encounters a static array parameter.  This is not compatible
with how GDC represents va_lists internally.

In the codegen pass, GDC will saturate a va_list to a pointer if its
underlying type is a static array.